+++ lib/osf-components/addon/components/delete-button/component.ts
@@ -4,6 +4,7 @@ import { service } from '@ember-decorators/service';
 import Component from '@ember/component';
 import { task } from 'ember-concurrency';
 import I18n from 'ember-i18n/services/i18n';
+import Toast from 'ember-toastr/services/toast';
 
 import requiredAction from 'ember-osf-web/decorators/required-action';
 import Analytics from 'ember-osf-web/services/analytics';
@@ -16,15 +17,21 @@ import layout from './template';
 @tagName('span')
 export default class DeleteButton extends Component.extend({
     _deleteTask: task(function *(this: DeleteButton) {
-        if (this.analyticsScope) {
-            this.analytics.click('button', `${this.analyticsScope} - Confirm delete`);
+        try {
+            if (this.analyticsScope) {
+                this.analytics.click('button', `${this.analyticsScope} - Confirm delete`);
+            }
+            yield this.delete();
+            this.set('modalShown', false);
+        } catch (e) {
+            this.toast.error(this.errorMessage);
+            throw e;
         }
-        yield this.delete();
-        this.set('modalShown', false);
     }).drop(),
 }) {
     @service analytics!: Analytics;
     @service i18n!: I18n;
+    @service toast!: Toast;
 
     // Required arguments
     @requiredAction delete!: () => unknown;
@@ -53,6 +60,10 @@ export default class DeleteButton extends Component.extend({
     cancelButtonText: string = defaultTo(
         this.cancelButtonText,
         this.i18n.t('osf-components.delete-button.cancelButtonText'),
+    );
+    errorMessage: string = defaultTo(
+        this.errorMessage,
+        this.i18n.t('osf-components.delete-button.error'),
     );
 
     // Private properties